# src/shared/infra/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TypeOrm_Postgre API",
    "description": "...",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/",
  "tags": [
    {
      "name": "Authenticated Routes",
      "description": "Routes of Class Users",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "Users_auth",
      "description": "Endpoints for adding pets to the directory"
    },
    {
      "name": "Products_auth",
      "description": "Endpoints for updating pet information"
    },
    {
      "name": "Categories_auth",
      "description": "Endpoints for getting information about pets"
    },
    {
      "name": "Open Routes",
      "description": "Routes of Class Users",
      "externalDocs": {
        "description": "Find out more",
        "url": "http://swagger.io"
      }
    },
    {
      "name": "Users",
      "description": "Endpoints for adding pets to the directory"
    },
    {
      "name": "Products",
      "description": "Endpoints for updating pet information"
    },
    {
      "name": "Categories",
      "description": "Endpoints for getting information about pets"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/categories": {
      "post": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/products": {
      "post": {
        "tags": [
          "Products_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/changeProduct/{id}": {
      "put": {
        "tags": [
          "Products_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/allUsers": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/roles": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/users/search/{username}": {
      "get": {
        "tags": [
          "Users_auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/allProducts": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/allCategories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "User username."
          }
        ],
        "responses": {}
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/auth/signin": {
      "post": {
        "tags": [
          "Users"
        ],
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "mario-lucca@hotmail.com"
                },
                "password": {
                  "type": "string",
                  "example": "1234"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  }
}